Add math label for bunching
